syntax/logic fixes for multi-/boot support (still beta) and font changes
The font changes are designed to allow grub to install fonts to
/usr/share/grub/fonts and have them copied to /boot/grub if they are
not there already. This eliminates the need for the grub ebuild to
mount /boot when it is installed for the purposes of installing the
included unifont.pf2 font.

The multi-/boot support that was very rough at least should no longer
prevent boot-update from running. non-multi-/boot support now works.
(Multi-/boot support is the new, beta ability to specify a non /boot
location to scan for kernels and initrds.)